cwi-engaja/api/src/main/java/br/com/cwi/crescer/api/service/BuscarUsuarioService.java
package br.com.cwi.crescer.api.service;

import br.com.cwi.crescer.api.domain.Usuario;
import br.com.cwi.crescer.api.repository.UsuarioRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.server.ResponseStatusException;

@Service
public class BuscarUsuarioService {
    @Autowired
    UsuarioRepository usuarioRepository;

    public Usuario buscar(String usuario) {
        return usuarioRepository
                .findByUser(usuario)
                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST,
                        "Houve um problema ao buscar informações do usuário."));
    }
}
